+++ opentelemetry-sdk/src/metrics/mod.rs
@@ -49,10 +49,10 @@ pub(crate) mod manual_reader;
 pub(crate) mod meter;
 mod meter_provider;
 pub(crate) mod noop;
-#[cfg(not(feature = "experimental_metrics_periodicreader_with_async_runtime"))]
 pub(crate) mod periodic_reader;
 #[cfg(feature = "experimental_metrics_periodicreader_with_async_runtime")]
-pub(crate) mod periodic_reader_with_async_runtime;
+/// Module for periodic reader with async runtime.
+pub mod periodic_reader_with_async_runtime;
 pub(crate) mod pipeline;
 pub mod reader;
 pub(crate) mod view;
@@ -61,10 +61,7 @@ pub use aggregation::*;
 pub use error::{MetricError, MetricResult};
 pub use manual_reader::*;
 pub use meter_provider::*;
-#[cfg(not(feature = "experimental_metrics_periodicreader_with_async_runtime"))]
 pub use periodic_reader::*;
-#[cfg(feature = "experimental_metrics_periodicreader_with_async_runtime")]
-pub use periodic_reader_with_async_runtime::*;
 pub use pipeline::Pipeline;
 
 pub use instrument::InstrumentKind;
@@ -515,9 +512,6 @@ mod tests {
         }
 
         let exporter = InMemoryMetricExporter::default();
-        #[cfg(feature = "experimental_metrics_periodicreader_with_async_runtime")]
-        let reader = PeriodicReader::builder(exporter.clone(), runtime::Tokio).build();
-        #[cfg(not(feature = "experimental_metrics_periodicreader_with_async_runtime"))]
         let reader = PeriodicReader::builder(exporter.clone()).build();
         let meter_provider = SdkMeterProvider::builder().with_reader(reader).build();
 
@@ -534,9 +528,6 @@ mod tests {
     async fn counter_duplicate_instrument_merge() {
         // Arrange
         let exporter = InMemoryMetricExporter::default();
-        #[cfg(feature = "experimental_metrics_periodicreader_with_async_runtime")]
-        let reader = PeriodicReader::builder(exporter.clone(), runtime::Tokio).build();
-        #[cfg(not(feature = "experimental_metrics_periodicreader_with_async_runtime"))]
         let reader = PeriodicReader::builder(exporter.clone()).build();
         let meter_provider = SdkMeterProvider::builder().with_reader(reader).build();
 
@@ -588,9 +579,6 @@ mod tests {
     async fn counter_duplicate_instrument_different_meter_no_merge() {
         // Arrange
         let exporter = InMemoryMetricExporter::default();
-        #[cfg(feature = "experimental_metrics_periodicreader_with_async_runtime")]
-        let reader = PeriodicReader::builder(exporter.clone(), runtime::Tokio).build();
-        #[cfg(not(feature = "experimental_metrics_periodicreader_with_async_runtime"))]
         let reader = PeriodicReader::builder(exporter.clone()).build();
         let meter_provider = SdkMeterProvider::builder().with_reader(reader).build();
 
@@ -680,9 +668,6 @@ mod tests {
     async fn instrumentation_scope_identity_test() {
         // Arrange
         let exporter = InMemoryMetricExporter::default();
-        #[cfg(feature = "experimental_metrics_periodicreader_with_async_runtime")]
-        let reader = PeriodicReader::builder(exporter.clone(), runtime::Tokio).build();
-        #[cfg(not(feature = "experimental_metrics_periodicreader_with_async_runtime"))]
         let reader = PeriodicReader::builder(exporter.clone()).build();
         let meter_provider = SdkMeterProvider::builder().with_reader(reader).build();
 
@@ -820,9 +805,6 @@ mod tests {
 
         // Arrange
         let exporter = InMemoryMetricExporter::default();
-        #[cfg(feature = "experimental_metrics_periodicreader_with_async_runtime")]
-        let reader = PeriodicReader::builder(exporter.clone(), runtime::Tokio).build();
-        #[cfg(not(feature = "experimental_metrics_periodicreader_with_async_runtime"))]
         let reader = PeriodicReader::builder(exporter.clone()).build();
         let criteria = Instrument::new().name("my_observable_counter");
         // View drops all attributes.
@@ -898,9 +880,6 @@ mod tests {
 
         // Arrange
         let exporter = InMemoryMetricExporter::default();
-        #[cfg(feature = "experimental_metrics_periodicreader_with_async_runtime")]
-        let reader = PeriodicReader::builder(exporter.clone(), runtime::Tokio).build();
-        #[cfg(not(feature = "experimental_metrics_periodicreader_with_async_runtime"))]
         let reader = PeriodicReader::builder(exporter.clone()).build();
         let criteria = Instrument::new().name("my_counter");
         // View drops all attributes.